# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cyber Code API",
    "version": "1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4747"
    }
  ],
  "security": [
    {
      "basic": []
    }
  ],
  "components": {
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "LocationInfo": {
        "type": "object",
        "properties": {
          "directory": {
            "type": "string"
          },
          "workspace": {
            "type": [
              "string",
              "null"
            ]
          },
          "project": {
            "$ref": "#/components/schemas/ProjectInfo"
          }
        },
        "required": [
          "directory",
          "project"
        ]
      },
      "ProjectInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "directory": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "directory"
        ]
      },
      "Located_Page_SessionRow": {
        "type": "object",
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LocationInfo"
          },
          "data": {
            "$ref": "#/components/schemas/Page_SessionRow"
          }
        },
        "required": [
          "location",
          "data"
        ],
        "description": "A Location-scoped response."
      },
      "Page_SessionRow": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionRow"
            }
          },
          "cursor": {
            "$ref": "#/components/schemas/Cursor"
          }
        },
        "required": [
          "data",
          "cursor"
        ]
      },
      "SessionRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "directory": {
            "type": "string"
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          },
          "cost": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "id",
          "title",
          "directory",
          "model",
          "archived",
          "created_at",
          "updated_at",
          "cost"
        ]
      },
      "Cursor": {
        "type": "object",
        "properties": {
          "previous": {
            "type": [
              "string",
              "null"
            ]
          },
          "next": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "CreateBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ],
            "description": "`provider/model`; defaults to the configured `model`."
          },
          "agent": {
            "type": [
              "string",
              "null"
            ]
          },
          "mode": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "rules": {
            "description": "Session ruleset in the `permissions` config shape."
          },
          "max_steps": {
            "type": [
              "integer",
              "null"
            ],
            "format": "uint32",
            "minimum": 0
          }
        }
      },
      "Located_Session": {
        "type": "object",
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LocationInfo"
          },
          "data": {
            "$ref": "#/components/schemas/Session"
          }
        },
        "required": [
          "location",
          "data"
        ],
        "description": "A Location-scoped response."
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "directory": {
            "type": "string"
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "agent": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "created_ms": {
            "type": "integer",
            "format": "int64"
          },
          "archived": {
            "type": "boolean",
            "default": false
          },
          "default_title": {
            "type": "boolean",
            "description": "The title is still the generated default.",
            "default": false
          },
          "rules": {
            "description": "Session ruleset (`permissions-modes`), in the `permissions` config shape."
          },
          "max_steps": {
            "type": [
              "integer",
              "null"
            ],
            "format": "uint32",
            "minimum": 0,
            "description": "Step limit for this Session's Drains, on top of the runtime-wide limit."
          },
          "status": {
            "type": "string",
            "description": "`idle` or `running`."
          },
          "seq": {
            "type": "integer",
            "format": "int64",
            "description": "Sequence of the last durable event; stream from here to follow new activity."
          },
          "totals": {
            "$ref": "#/components/schemas/Totals"
          },
          "revert": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RevertState"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "title",
          "directory",
          "agent",
          "model",
          "mode",
          "created_ms",
          "status",
          "seq",
          "totals"
        ],
        "description": "A Session with its live status."
      },
      "Totals": {
        "type": "object",
        "properties": {
          "usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "cost": {
            "type": "number",
            "format": "double"
          },
          "unpriced_steps": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0,
            "description": "Steps whose price was unknown (`unpriced`), so `cost` is a lower bound."
          },
          "steps": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          }
        },
        "required": [
          "usage",
          "cost",
          "unpriced_steps",
          "steps"
        ]
      },
      "Usage": {
        "type": "object",
        "properties": {
          "input": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "output": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "reasoning": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "cache_read": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "cache_write": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          }
        },
        "required": [
          "input",
          "output",
          "reasoning",
          "cache_read",
          "cache_write"
        ],
        "description": "Token classes of one step. `input` excludes cache reads and writes; `output` excludes\nreasoning."
      },
      "RevertState": {
        "type": "object",
        "properties": {
          "message_id": {
            "type": "string",
            "description": "The user message rewound to; it and everything after it go on commit."
          },
          "target": {
            "$ref": "#/components/schemas/RevertTarget"
          },
          "baseline": {
            "type": [
              "string",
              "null"
            ],
            "description": "The working tree before the first stage."
          },
          "applied": {
            "type": [
              "string",
              "null"
            ],
            "description": "The tree the working tree was restored to."
          },
          "diff": {
            "type": "string",
            "description": "Unified diff from the baseline to the restored tree."
          }
        },
        "required": [
          "message_id",
          "target",
          "diff"
        ],
        "description": "A staged revert: the conversation boundary and the trees involved."
      },
      "RevertTarget": {
        "type": "string",
        "enum": [
          "code",
          "conversation",
          "both"
        ],
        "description": "What a rewind restores (`snapshots-checkpoints` → Three-phase revert)."
      },
...
      "EventEnvelope": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "data": true,
          "location": {
            "type": [
              "string",
              "null"
            ],
            "description": "The Location directory the event belongs to."
          },
          "durable": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DurableRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "type",
          "data"
        ],
        "description": "One event on any stream."
      },
      "DurableRef": {
        "type": "object",
        "properties": {
          "aggregateID": {
            "type": "string"
          },
          "seq": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          }
        },
        "required": [
          "aggregateID",
          "seq",
          "version"
        ],
        "description": "Where a durable event sits in its aggregate's history."
      },
...
      "ErrorBody": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "title": "ErrorBody",
        "description": "The JSON body of every declared failure.",
        "type": "object",
        "properties": {
          "_tag": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "ref": {
            "description": "`err_<id>` for `UnknownError`, logged server-side.",
            "type": [
              "string",
              "null"
            ]
          },
          "paths": {
            "description": "Paths for rewind conflicts.",
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "_tag",
          "message"
        ]
      }
    }
  },
  "paths": {
...
    "/api/v1/openapi.json": {
      "get": {
        "operationId": "v1.health.openapi",
        "summary": "This document",
        "tags": [
          "health"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ws": {
      "get": {
        "operationId": "v1.event.websocket",
        "summary": "JSON-RPC 2.0 over WebSocket: operation IDs as methods, events as notifications",
        "tags": [
          "event"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "x-websocket": true
      }
    },
    "/api/v1/event": {
      "get": {
        "operationId": "v1.event.subscribe",
        "summary": "Live events for a Location (scope=all for every Location)",
        "tags": [
          "event"
        ],
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location[directory]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cyber-directory",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events; each `data` is an EventEnvelope",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "get": {
        "operationId": "v1.session.list",
        "summary": "List root Sessions of a Location, newest first",
        "tags": [
          "session"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "children",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location[directory]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cyber-directory",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Located_Page_SessionRow"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "v1.session.create",
        "summary": "Create a Session",
        "tags": [
          "session"
        ],
        "parameters": [
          {
            "name": "location[directory]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cyber-directory",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Located_Session"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBody"
              }
            }
          }
        }
      }
    },
...
  }
}